Modifeid logic for thread count

+++ Ramses/ramses.py
@@ -51,7 +51,7 @@ class RamsesMPI_strong(RamsesMPI):
     def set_job_script_variables(self):
 
         if self.current_partition.processor.num_cpus_per_core > 1:
-            self.core_count_1_node = self.current_partition.processor.num_cpus/self.current_partition.processor.num_cpus_per_core
+            self.core_count_1_node = int(self.current_partition.processor.num_cpus/self.current_partition.processor.num_cpus_per_core)
         else:    
             self.core_count_1_node = self.current_partition.processor.num_cpus 
         
@@ -79,7 +79,7 @@ class RamsesMPI_weak(RamsesMPI):
     def set_job_script_variables(self):
 
         if self.current_partition.processor.num_cpus_per_core > 1:
-            self.core_count_1_node = self.current_partition.processor.num_cpus/self.current_partition.processor.num_cpus_per_core
+            self.core_count_1_node = int(self.current_partition.processor.num_cpus/self.current_partition.processor.num_cpus_per_core)
         else:    
             self.core_count_1_node = self.current_partition.processor.num_cpus
